move-js-export: js.ErgoTree moved to interpreter module
