+++ sdk/js/src/main/scala/org/ergoplatform/sdk/js/Isos.scala
@@ -32,18 +32,6 @@ import scala.scalajs.js.Object
 
 /** Definitions of isomorphisms. */
 object Isos {
-  /** Conversion between `Value` and `Constant[SType]`. */
-  implicit val isoValueToConstant: Iso[Value, Constant[SType]] = new Iso[Value, Constant[SType]] {
-    override def to(x: Value): Constant[SType] =
-      Constant(x.runtimeData.asInstanceOf[SType#WrappedType], Evaluation.rtypeToSType(x.tpe.rtype))
-
-    override def from(x: Constant[SType]): Value = {
-      val rtype = Evaluation.stypeToRType(x.tpe)
-      val jsvalue = Value.fromRuntimeData(x.value, rtype)
-      new Value(jsvalue, new Type(rtype))
-    }
-  }
-
   val isoStringToGroupElement: Iso[String, sigma.GroupElement] = new Iso[String, sigma.GroupElement] {
     override def to(x: String): sigma.GroupElement = {
       val bytes = Base16.decode(x).get
